Split functions in several files
